fix(YALB-1137): div now matches storybook
This is an attempt to make the dials part of the component over being
separated from it.  This allows us to set paramters to define the width
should we decide we want to.

I am awaiting more feedback for other ways to do this, but wanted to get
this recorded so we can make changes later if needed.