feat(YALB-1137): position/width now in component
The position and width are now used inside of the twig component.

The position is placed on the inner div due to it being responsible for
the overall size of the container, while the div handles its thickness
and width within that container.

The size is placed on the divider div itself.

Due to what I saw as a possible semantic issue between divider__width as
it used to be, which didn't seem to be defining the width of the
divider, but an offset of some kind, I renamed the functionality of that
to divider__component_width since that is what it references in the CSS
ultimately.  This allowed us to reuse divider__width to define the
percentage width of the divider itself.