fix(YALB-1195): format now displays on string and array
The component was not rendering the format unless it was an array type,
although other pieces like storybook had a default that was a string.
This change allows it to reference the 0th element if it's a single
element array, or the string text if it was passed.

My only question is: was this not done because an empty string truthy?
So if they pass an empty string, would it think there was a format and
display it?