chore(YALB-1042): document percy auto-approve
Percy was not really being utilized right now, so in an event to keep
the workflow for use in the future, changes have been made in the
Percy.io site project settings that instead of only auto-approving
the main branch, we auto-approve all branches.

When reverting this, just change it back to main.