fix(YALB-1197): create fallback height
For those embeds that might not contain a data-embed-type attribute, we
want it to revert to the unknown variant.  Examples of this is the
instagram embed, that converts itself on execution of the iframe, not by
us.  This allows it to revert back to the default that existed before
this new implementation idea.