revert(YALB-1197): do not have fallback height
After speaking to @myvaughn, Instagram adhering to a 70vh, yet still
displaying unscrollable content is an issue for keyboard users who would
still be able to tab into the content, while mouse users would not.  As
such, the decision was made to allow those not targeted with the
embed_type to handle their height as they see fit.

As a result, Instagram will now display comments underneat the image
displayed in browsers that can convert the blockquote to an iframe.