fix(YALB-1273): rearrange example to show height
The original YAML file had the long text at the end of the file,
resulting on a 1080p monitor as showing on a separate line all by
itself, and not featuring the new implementation of the height of a
button in a row being set to that of the largest link.  By placing it in
the number three spot, it now better shows this off.

I also added the word 'very' to make it even longer of a text entry.