fix(YALB-1418): use unix timestamp for 'now' eval
While the start and end dates where being passed in as unix timestamps,
the current date/time was being converted to a human readable date,
causing unexpected behavior in the determination of whether the event
has already passed or not.

This change converts the evaluation of the event passing with the unix
timestamp for the current date and time.