feat(YALB-1445): graceful fallback for accordion controls
While the existing JavaScript gracefully fell back to an inital state of
the accordion items to be expanded, it allowed the controls to remain,
even though they are useless at that point.  This feature does a similar
thing by initially hiding the controls, and only showing them if
JavaScript is enabled.

The result for those without JavaScript is an expanded list of items
without "Expand/Collapse" controls.