fix(YALB-1394): make storybook update javascript
When changing some values in storybook, it does not re-attach the
javascript associated with items.  This has resulted in changing items,
traveling off of the control, and returning so taht the javascript gets
re-attached and ran.  Since it creates a new context object each time
values are changed, setting this to the useEffect will allow the drupal
behaviors to re-attach.