fix(YALB-1410): contrast passes on action-banner and grand hero H2
Action banner and grand hero H2 elements were causing WAVE issues when
scanned.  This change fixes this by first moving the theme attribute to
the top of the component so it can take advantage of the colors used for
the theme throughout, and then creating a background behind the
component using the color theme already used as the background in the
before pseudo class that should hardly ever be seen to allow WAVE and
other tools to pass.