fix(YALB-520): top level mega menu items not capitalized
I do not like admitting this, but the blame showed I added this CSS, and
I can't remember doing it.  In any event, it caused the mega menu top
level links with no submenu to display as uppercase, which looked funny.
I've removed this CSS and now all is well.