fix(YALB-1667): basic menu uses parent node title (#320)
The desired menu functionality for a basic menu is that the top level
menu should be the parent menu link title, and the first submenu item be
the parent page title.  This commit allows this functionality to happen.
In the event the top level menu is not associated with a node, it
reverts back to the parent menu link title.