fix(YALB-520): siteAnimationEnabled defaults to false
The siteAnimationEnabled is used as a way to determine if observers
should be used when the animation is set to something other than default
(for now, artistic).  As such, we can only assume that if for some
reason it does not exist, it SHOULD be the default, which would result
that evaluation to a false instead of true.  This reflects that.