fix(YALB-520): fix extra space in link text
For some reason, in things like Call to action buttons, there was an
extra space, making links look strange when you hovered over them.  By
adding the following to remove spaces in certain places, and eliminiate
the newline between the twig content and the icon defintion, the space
is now gone.