YALB-520: Links: Dynamically apply link treatments (#319)
* feat(YALB-520): add link decorations

While our component library has link decorations defined, we do not use
them throughout the component library.  This adds twig-based decorations
to existing components, as well as a javascript decorator to handle any
links that could not be addressed via twig, like links inside of a
CKEditor output.

* fix(YALB-520): alert link no longer uses visited color

Using the visited link color resulted in a contrast issue with critical
alerts.  We revert this back to using the regular alert color text until
we can decide a new color if needed.

* fix(YALB-520): fix extra space in link text

For some reason, in things like Call to action buttons, there was an
extra space, making links look strange when you hovered over them.  By
adding the following to remove spaces in certain places, and eliminate
the newline between the twig content and the icon definition, the space
is now gone.