fix(YSP-353): override white footer link shadow
When changing to a background that is not white, the links that are
decorated with underlines have an attribute of a white shadow on it.
This results in a terrible looking link.  The way this passed in the
past was that we had a white background with a white shadow, so it
looked ok.  This now mimics the same attempting to match the footer
background.  Right now we are just targeting the footer since it was
found only here, but it makes me wonder if we need the shadow?