feat(YSP-313): support 'loaded' mobile menu (#340)
* feat(YSP-313): support 'loaded' mobile menu

A loaded menu state is one where the page has been loaded, but no
toggles have been made.  This lets us take that into account and not
display the visual rollup on a resize.

* fix(YSP-313): fix arrows showing on resize

The svg arrows were showing even though the menu was not.  They no
longer show.

* feat(YSP-313): reset to loaded if closed

If we resize the window again and the current state of the menu is
closed, make it loaded so we don't get the roll up effect afterward.  If
it's open, leave it as is.

---------

Co-authored-by: Jim Vomero <jim.vomero@fourkitchens.com>