fix(YSP-426): skip links with image decoration
The mega footer allows images to be linked, which could be external,
causing the external link to occur in some cases, which looks odd.  This
skips those and any that would contain an image inside of the link since
it would look odd in most cases I can think of.