feat(YSP-000): quick-links cta display under
When changing to using a span solution for the link treatment, this
caused only the quick links to behave in a way that wasn't expected,
which is that the horizontal flex caused the span to display to the
right of the rest of the text.  This change ensures that it wraps
underneath, which looks more in line with what users expect.