fix(YSP-000): attempt to fix quick links button
Instead of treating this inside of link treatment, this is an attempt to
fix the issue at the component level.

This attempts to fix flex issues with the span on Safari since at some
screen sizes, it will cause strange wrapping that doesn't look good.
This allows it to display in an inline manner with the rest of the link
text by wrapping the whole thing.