YSP-569: V1.4.0 Spotlight in banner area has extra padding above (#382)
* fix(YSP-569): fix spacing of spotlights in banners

We originally had javascript that would decorate certain components in
order to allow us to modify the margin above and below to give a little
space between them.  In the context of adding these into a bannher,
however, we do not want this spacing for the first element as it creates
a space in between the header and the section.  This allows us to
target those specifically inside of the banner and remove this top
margin so that it bumps up to the header.

* feat(YSP-569): show banner spotlight in storybook