fix(YSP-559): fill cost_button_text with ctaText
Originally, the ctaText would not populate the cost_button_text, which
drives the CTA text for the button next to "Add to Calendar", resulting
in an empty button with decoration.  This allows it to be for both the
cta_primary_content and the cost_button_text so that if the other
existed, it'll use it there as well.