fix(YSP-485): patch @drupal/once package.json
With more recent node versions, you must specify your exclude
definitions with a relative path using the following: ./

There is work being done to bring this into the project, but it has not
been accepted yet.  Once that is accepted and a version is cut, we can
remove this patch.

References: https://www.drupal.org/project/once/issues/3466026