feat(YSP-680): mimic grand-hero for image-banner
Re-use grand hero to provide an identical experience.

Note that a possible improvement would be to utilize the grand hero
itself over a duplicated image-banner since without a lot of the
component text/link content, it should display the same.  I decided to
not go this route so that our markup reflects that it truely is an
image banner and not just a grand hero.