feat(YSP-480): determine data attribute dynamically
Since we are dealing with two different component names for data
attribute assigning, we want to select the correct one as we are looping
to name the attribute correctly.  This refactors a little bit of code
such that we can reuse it in the evaluation of the element to determine
the right name to use.