feat(YSP-762): refactor date generation with override
The date logic was quite complex.  We refactored this so that we could
easily add more date formats and reference them with an override if we
needed a specific one.  An example of this was for the post
"By author | DATE", which was formatted outside of this twig.  Now we
can include an override for items like this to force a certain style,
while allowing the existing functionality to remain.