fix(YSP-883): reuse global themes in tokens for lever
Before now, the themes available in storybook for selection in the
toolbar were hard coded, meaning we had to remember to add it here when
any new theme was added.  This now does similar to the `Color Global
Themes` section where it pulls this data from the tokens.json file.
This will allow less work in the future to keep this up to date.