test(card): add contrast testing story for reference card
Added a Storybook story to display all 30 theme combinations of the
reference card for accessibility contrast analysis. Enables a11y addon
to automatically flag any contrast violations.

This is being added to later re-use in a different way.  This will be
reverted immediately, but it will be here for when we need it again.